fix: fix bugs

+++ backend/participate/views.py
@@ -11,29 +11,26 @@ blueprint = Blueprint('participate', __name__)
 
 class ParticipateResource(Resource):
     def get(self):
-        form = request.get_json(True, True)
-        user_id = form.get('user_id')
-        task_id = form.get('task_id')
-        status = form.get('status')
+        user_id = request.args.get('user_id')
+        task_id = request.args.get('task_id')
+        status = request.args.get('status')
         participates = Participate.get(user_id=user_id, task_id=task_id, status=status)
         result = [{"id": participate.id,
                    "user_id": participate.user_id,
                    "task_id": participate.task_id,
-                   "status": ParticipateStatusCN[Participate.status]} for participate in participates]
+                   "status": ParticipateStatusCN[participate.status]} for participate in participates]
         return dict(data=result), 200
 
     def post(self):
         user_id = auth_helper()
         form = request.get_json(True, True)
         task_id = form.get('task_id')
-        participator_id = form.get('participator_id')
-        if not participator_id:
-            return dict(error='参与者不能为空'), 400
         if not task_id:
             return dict(error='请指定任务'), 400
         task = Task.get(task_id=task_id)
         if not task:
-            return dict(error='该任务不存在'), 400
+            return dict(error='该任务不存在'), 400  # FIXME 自己的任务自己能否参与
+        task = task[0]
         if not task.status:
             return dict(error='该任务尚未发布'), 400
         try:
@@ -63,12 +60,14 @@ class ParticipateResource(Resource):
         task = task[0]
         participator_id = form.get('participator_id')
         if not participator_id:
-            return dict(error="该参与者不存在"), 400
-        if task.start_time.strftime("%Y-%m-%d %H:%M") < get_cur_time():
-            return dict(error="任务已开始，不能进行人员调整"), 400
-        participate = Participate.get(user_id=user_id, task_id=task_id)
+            return dict(error="参与者不能为空"), 400
+        participate = Participate.get(user_id=participator_id, task_id=task_id)
         if not participate:
             return dict(error="该参与信息不存在"), 400
+        if user_id != participator_id and user_id != participator_id:  # 自己或者任务发布者才能退出任务或请人离开任务
+            return dict(error="您没有操作权限"), 403
+        if task.start_time.strftime("%Y-%m-%d %H:%M") < get_cur_time():
+            return dict(error="任务已开始，任务人员无法调整"), 400
         participate = participate[0]
         db.session.delete(participate)
         db.session.commit()
@@ -79,7 +78,7 @@ class ParticipateResource(Resource):
         return dict(data="ok"), 200
 
 
-@blueprint.route('/permit_application', methods=['POST'])
+@blueprint.route('/permit', methods=['POST'])
 def permit_application():  # 甲方审批(通过)
     form = request.get_json(True, True)
     user_id = auth_helper()
